jobportalbackend/requirements.txt
fastapi
uvicorn
bcrypt
mysql-connector-python
pydantic
Employee
passlib
